+++ src/eca/server.clj
@@ -115,6 +115,9 @@
 (defmethod jsonrpc.server/receive-request "chat/fork" [_ components params]
   (eventually (handlers/chat-fork (with-config components) params)))
 
+(defmethod jsonrpc.server/receive-notification "chat/update" [_ components params]
+  (async-notify (handlers/chat-update (with-config components) params)))
+
 (defmethod jsonrpc.server/receive-notification "mcp/stopServer" [_ components params]
   (async-notify (handlers/mcp-stop-server (with-config components) params)))